Unify all variants of the timeout argument for futex syscall
The timeout argument for a futex syscall--depending on the specific commands
and flags--may be interpreted as a relatiive or absolute time, and measured to
the CLOCK_MONOTONIC or CLOCK_REALTIME clock. To meet the various demands for
timeout, the FutexTimeout type is introduced. But this solution is not as neat
as it should be and mistaken the relativity of timeouts in some cases
(FUTEX_WAIT_BITSET takes absolute timeouts, not relative ones).

This commit unifies all the various cases of futex timeouts by converting the
timeout in a Duration that is always interpreted as a relative time. Handling
this way is simpler, more robust and elegant.