[libos] fix user stack content being overwritten
The current implementation of __occlum_sysret overwrites the contents of the user stack, which may causes crashes in leaf functions which depend on red zone. This commit moves the check for pku_enabled ahead of popfq to avoid the problem.

However, this problem still exists in the update_pkru_in_sysret function (since several registers inevitably need to be backed up before calling wrpkru). This commit provides a workaround by skipping red zone at the top of the stack.

Signed-off-by: Kun Lai <me@imlk.top>